Implementar modulo de categoria CRUD completo:
Implementar listar con id, filtro ,crear, eliminar,actualizar , cambiar estado y eliminar
ademas de una fix con los modulos afectados(ventas)

+++ src/main/java/com/sistemabarberia/fadex_backend/modules/servicio/service/ICorteService.java
@@ -11,11 +11,11 @@ public interface ICorteService {
 
     List<CorteResponseDTO> listar();
 
-    List<CorteResponseDTO> listarPorCategoria(Integer categoriaId);
+    List<CorteResponseDTO> listarPorCategoria(Long categoriaId);
 
-    CorteResponseDTO obtenerPorId(Integer id);
+    CorteResponseDTO obtenerPorId(Long id);
 
-    CorteResponseDTO actualizar(Integer id, CorteRequestDTO dto);
+    CorteResponseDTO actualizar(Long id, CorteRequestDTO dto);
 
-    void eliminar(Integer id);
+    void eliminar(Long id);
 }
